@@ -4,12 +4,11 @@ import android.support.test.espresso.NoMatchingViewException;
 import android.support.test.espresso.UiController;
 import android.support.test.espresso.ViewAction;
 import android.support.test.espresso.ViewAssertion;
-import android.support.v4.view.ViewPager;
+import androidx.core.view.ViewPager;
 import android.view.View;
 
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
-import org.hamcrest.core.IsAnything;
 
 import static org.hamcrest.Matchers.anything;
 
