@@ -1,12 +1,14 @@
 package com.bassem.lastfm.adapters;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.bassem.lastfm.R;
 import com.bassem.lastfm.models.Album;
